+++ src/main/java/org/woehlke/twitterwall/conf/WebMvcConfig.java
@@ -1,0 +1,18 @@
+package org.woehlke.twitterwall.conf;
+
+
+import org.springframework.context.annotation.Configuration;
+import org.springframework.web.servlet.config.annotation.ViewControllerRegistry;
+import org.springframework.web.servlet.config.annotation.WebMvcConfigurerAdapter;
+
+@Configuration
+public class WebMvcConfig extends WebMvcConfigurerAdapter {
+
+    @Override
+    public void addViewControllers(ViewControllerRegistry registry) {
+        registry.addViewController("/adm").setViewName("redirect:/application/management");
+        registry.addViewController("/").setViewName("redirect:/tweet/all");
+    }
+
+}
+
