+++ src/main/java/org/woehlke/twitterwall/frontend/controller/ApplicationController.java
@@ -4,7 +4,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.RequestMapping;
-import org.woehlke.twitterwall.conf.TwitterProperties;
+import org.woehlke.twitterwall.conf.properties.TwitterProperties;
 import org.woehlke.twitterwall.frontend.controller.common.Symbols;
 import org.woehlke.twitterwall.frontend.controller.common.ControllerHelper;
 import org.woehlke.twitterwall.oodm.entities.parts.CountedEntities;
@@ -38,7 +38,7 @@ public class ApplicationController {
         model = controllerHelper.setupPage(model,title,subtitle,symbol);
         return "application/management";
     }
-    
+
     @Autowired
     public ApplicationController(CountedEntitiesService countedEntitiesService, ControllerHelper controllerHelper, TwitterProperties twitterProperties) {
         this.countedEntitiesService = countedEntitiesService;
